requirements.txt
asyncio>=3.4.3
dataclasses>=0.6
pytest>=5.3.5
rdflib>=4.2.2
rdflib-jsonld>=0.4.0
requests>=2.22.0
SPARQLWrapper>=1.8.5
websockets>=8.1
python-dotenv>=0.13.0
trompace-client
